update to build on osx (intel)

+++ src/ReadHelper.h
@@ -44,7 +44,7 @@ void findSyncmerPositions(const SequenceCharType& sequence, size_t kmerSize, siz
 		size_t seqPos = smerSize+i-1;
 		fwkmerHasher.addChar(sequence[seqPos]);
 		fwkmerHasher.removeChar(sequence[seqPos-smerSize]);
-		size_t hash = fwkmerHasher.hash();
+		uint64_t hash = fwkmerHasher.hash();
 		if (endSmer(hash)) hash = 0;
 		while (smerOrder.size() > 0 && std::get<1>(smerOrder.back()) > hash) smerOrder.pop_back();
 		smerOrder.emplace_back(i, hash);
@@ -58,7 +58,7 @@ void findSyncmerPositions(const SequenceCharType& sequence, size_t kmerSize, siz
 		size_t seqPos = smerSize+i-1;
 		fwkmerHasher.addChar(sequence[seqPos]);
 		fwkmerHasher.removeChar(sequence[seqPos-smerSize]);
-		size_t hash = fwkmerHasher.hash();
+		uint64_t hash = fwkmerHasher.hash();
 		if (endSmer(hash)) hash = 0;
 		// even though pop_front is used it turns out std::vector is faster than std::deque ?!
 		// because pop_front is O(w), but it is only called in O(1/w) fraction of loops
